feat: Initial project setup and Info.plist fix; Update PROJECT_PLAN.md with checkboxes and GEMINI.md with self-correction instructions
