--- Flume App/Flume_AppApp.swift
@@ -1,32 +1,0 @@
-//
-//  Flume_AppApp.swift
-//  Flume App
-//
-//  Created by Sholto Maud on 25/6/2025.
-//
-
-import SwiftUI
-import SwiftData
-
-@main
-struct Flume_AppApp: App {
-    var sharedModelContainer: ModelContainer = {
-        let schema = Schema([
-            Item.self,
-        ])
-        let modelConfiguration = ModelConfiguration(schema: schema, isStoredInMemoryOnly: false)
-
-        do {
-            return try ModelContainer(for: schema, configurations: [modelConfiguration])
-        } catch {
-            fatalError("Could not create ModelContainer: \(error)")
-        }
-    }()
-
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-        .modelContainer(sharedModelContainer)
-    }
-}
