+++ FlumeApp/FlumeApp/Assets.xcassets/Contents.json
