+++ FlumeApp/FlumeApp/ContentView.swift
@@ -1,0 +1,24 @@
+//
+//  ContentView.swift
+//  FlumeApp
+//
+//  Created by Sholto Maud on 28/6/2025.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        VStack {
+            Image(systemName: "globe")
+                .imageScale(.large)
+                .foregroundStyle(.tint)
+            Text("Hello, world!")
+        }
+        .padding()
+    }
+}
+
+#Preview {
+    ContentView()
+}
