+++ FlumeApp/FlumeApp/FlumeAppApp.swift
@@ -1,0 +1,17 @@
+//
+//  FlumeAppApp.swift
+//  FlumeApp
+//
+//  Created by Sholto Maud on 28/6/2025.
+//
+
+import SwiftUI
+
+@main
+struct FlumeAppApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
